requirements.txt
-i https://pypi.org/simple
aiocache==0.11.1
aiohttp==3.6.2
apscheduler==3.6.3
async-timeout==3.0.1
asyncio==3.4.3
atlassian-python-api==1.16.0
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core[grpc]==1.22.0
google-auth==1.19.2
google-cloud-secret-manager==1.0.0
googleapis-common-protos[grpc]==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.30.0
idna==2.10
importlib-metadata==1.7.0; python_version < '3.8'
iniconfig==1.0.1
more-itertools==8.4.0
multidict==4.7.6
oauthlib==3.1.0
packaging==20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pluggy==0.13.1
protobuf==3.12.2
py==1.9.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==2.4.7
pytest-curl-report==0.5.4
pytest-httpserver==0.3.5
pytest==6.0.1
python-json-logger==0.1.11
pytz==2020.1
requests-oauthlib==1.3.0
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.3; python_version < '3.8'
tzlocal==2.1
urllib3==1.25.10
werkzeug==1.0.1
yarl==1.5.1
zipp==3.1.0; python_version >= '3.6'
